# Conflicts:
#	src/main/resources/templates/checkout.html
#	src/main/resources/templates/sign-up.html

+++ src/main/java/com/ecommerce/g58/dto/OrdersDTO.java
@@ -12,6 +12,6 @@ public class OrdersDTO {
     private int orderId;
     private LocalDateTime orderDate;
     private String status;
-    private int totalQuantity;
-    private int totalPrice;
+    private BigDecimal totalQuantity;
+    private BigDecimal totalPrice;
 }
